Extracted CR Departures table into separate component.
