+++ assets/css/pre_fare_v2.scss
@@ -44,6 +44,7 @@
 @import "v2/pre_fare/cr_departures/cr_departures";
 @import "v2/pre_fare/cr_departures/overnight_cr_departures";
 @import "v2/pre_fare/cr_departures/cr_departures_header";
+@import "v2/pre_fare/cr_departures/cr_departures_table";
 
 @import "v2/pre_fare/subway_status";
 @import "v2/lcd_common_styles/route_pill";
