+++ assets/css/v2/pre_fare/cr_departures/cr_departures_table.scss
@@ -1,0 +1,96 @@
+.cr-departures-table {
+  border-collapse: collapse;
+  margin: 0 44px;
+  table-layout: fixed;
+  tr {
+    &:not(:first-child) {
+      height: 272px;
+    }
+    &:not(:last-child) {
+      border-bottom: 4px solid #cccbc8;
+    }
+  }
+  td {
+    padding: 30px 0;
+  }
+  .arrow {
+    width: 90px;
+  }
+  .headsign {
+    padding-left: 82px;
+    width: 516px;
+    font-size: 56px;
+    font-weight: bold;
+    letter-spacing: 0px;
+    line-height: 56px;
+  }
+  .arrival {
+    padding-left: 84px;
+    width: 180px;
+    height: 64px;
+    font-size: 72px;
+    font-weight: 800;
+    text-align: right;
+    line-height: 64px;
+    vertical-align: text-top;
+  }
+  .track,
+  .track-column {
+    width: 90px;
+    font-size: 56px;
+  }
+  .track-tbd {
+    width: 90px;
+    height: 90px;
+    text-align: center;
+    font-size: 36px;
+    font-weight: normal;
+    letter-spacing: 0px;
+    line-height: 32px;
+  }
+  .stops-at-text {
+    font-size: 40px;
+    font-weight: 600;
+    letter-spacing: 0px;
+    line-height: 40px;
+    margin-top: 16px;
+  }
+  .headsign-text {
+    width: 516px;
+    margin-bottom: 7px;
+  }
+  .track-number-text {
+    text-align: center;
+    height: 56px;
+  }
+  .arrow-image {
+    width: 90px;
+    height: 90px;
+    text-align: center;
+
+    .departure__arrow-image {
+      height: 64px;
+      width: 64px;
+    }
+  }
+  .via-service-icon {
+    display: inline-block;
+    vertical-align: middle;
+  }
+  .via-stop-name {
+    display: inline-block;
+    vertical-align: middle;
+    margin-left: 24px;
+  }
+
+  .table-header__english {
+    @include font--text--75bold;
+    font-size: 32px;
+    line-height: 44px;
+  }
+  .table-header__spanish {
+    @include font--text--55roman;
+    font-size: 24px;
+    line-height: 36px;
+  }
+}
